+++ src/main/java/dev/yanianz/essentials/guild/GuildRank.java
@@ -1,0 +1,5 @@
+package dev.yanianz.essentials.guild;
+
+public enum GuildRank {
+    MEMBER, OFFICER, LEADER
+}