+++ src/main/java/fr/maxlego08/essentials/commands/CommandLoader.java
@@ -6,6 +6,7 @@ import fr.maxlego08.essentials.api.commands.CommandManager;
 import dev.yanianz.essentials.nicknames.CommandNick;
 import dev.yanianz.essentials.networkchat.CommandGlobalChat;
 import dev.yanianz.essentials.friends.CommandFriend;
+import dev.yanianz.essentials.guild.CommandGuild;
 import dev.yanianz.essentials.reports.CommandReport;
 import dev.yanianz.essentials.reports.CommandReports;
 import dev.yanianz.essentials.notes.CommandNotes;
@@ -133,6 +134,7 @@ public class CommandLoader {
         register("tpr", CommandTeleportRandom.class, "rtp");
         register("g", CommandGlobalChat.class, "globalchat");
         register("friend", CommandFriend.class, "friends");
+        register("guild", CommandGuild.class, "g");
 
         register("balancetop", CommandBalanceTop.class, "baltop");
         register("economy", CommandEconomy.class, "eco");