+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -23,6 +23,7 @@ import dev.yanianz.essentials.discordbridge.DiscordBridgeModule;
 import dev.yanianz.essentials.network.BungeeChatModule;
 import dev.yanianz.essentials.networkchat.GlobalChatModule;
 import dev.yanianz.essentials.friends.FriendsModule;
+import dev.yanianz.essentials.guild.GuildModule;
 import dev.yanianz.essentials.chatcustomization.ChatCustomizationModule;
 import dev.yanianz.essentials.reports.ReportsModule;
 import dev.yanianz.essentials.notes.NotesModule;
@@ -119,6 +120,7 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(BungeeChatModule.class, new BungeeChatModule(this.plugin));
         this.modules.put(GlobalChatModule.class, new GlobalChatModule(this.plugin));
         this.modules.put(FriendsModule.class, new FriendsModule(this.plugin));
+        this.modules.put(GuildModule.class, new GuildModule(this.plugin));
         this.modules.put(DiscordBridgeModule.class, new DiscordBridgeModule(this.plugin));
         this.modules.put(SleepModule.class, new SleepModule(this.plugin));
         this.modules.put(TabListModule.class, new TabListModule(this.plugin));